requirements.txt
Django
django-cors-headers
djangorestframework
pandas
scikit-learn
seaborn
matplotlib